requirements.txt
requests>=2.31.0
openai-whisper>=20231117
torch>=2.0.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
